+++ crates/tools/cardano-tracer/src/acceptors/run.rs
@@ -311,6 +311,7 @@ mod tests {
             connected_nodes_names: ConnectedNodesNames::new(),
             accepted_metrics: crate::metrics_store::new_accepted_metrics(),
             handle_registry: crate::types::HandleRegistry::new(),
+            data_point_requestors: crate::types::DataPointRequestors::new(),
             network_magic: 764824073,
         }
     }